chore: remove javadocs (#21)
The Javadoc comments are currently not published in a machine readable
form and are difficult to parse 100% correctly from the published ADOC files.
Further, it might raise copyright-related questions. Further, the Java types
might slightly differ from the types described in the ADOC files. So, it seems
like a good idea to remove the Javadocs altogether.